pyproject.toml
[project]
name = "noor"
version = "0.1.0"
description = "AI-powered web navigator for visually impaired users"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.21.0",
    "google-cloud-aiplatform>=1.70.0",
    "google-cloud-firestore>=2.19.0",
    "google-cloud-secret-manager>=2.21.0",
    "google-cloud-logging>=3.11.0",
    "google-genai>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "playwright>=1.49.0",
    "pydantic>=2.10.0",
    "structlog>=24.4.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "Pillow>=11.0.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["noor_agent*", "server*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
